# ENH: Speed up sparse.csgraph.dijkstra