--- a/benchmarks/benchmarks/sparse_csgraph_dijkstra.py
+++ b/benchmarks/benchmarks/sparse_csgraph_dijkstra.py
@@ -20,7 +20,7 @@ class Dijkstra(Benchmark):
         rng = np.random.default_rng(1234)
         if format == 'random':
             # make a random connectivity matrix
-            data = scipy.sparse.rand(n, n, density=0.2, format='csc',
+            data = scipy.sparse.rand(n, n, density=0.2, format='lil',
                                      random_state=42, dtype=np.bool_)
             data.setdiag(np.zeros(n, dtype=np.bool_))
             self.data = data