--- a/scipy/sparse/csgraph/_shortest_path.pyx
+++ b/scipy/sparse/csgraph/_shortest_path.pyx
@@ -569,29 +569,37 @@ def dijkstra(csgraph, directed=True, indices=None,
         else:
             predecessor_matrix = np.empty((len(indices), N), dtype=ITYPE)
             predecessor_matrix.fill(NULL_IDX)
+            source_matrix = np.empty((len(indices), 0), dtype=ITYPE) # unused
     else:
         if min_only:
             predecessor_matrix = np.empty(0, dtype=ITYPE)
-            source_matrix = np.empty(0, dtype=ITYPE)
+            source_matrix = np.empty(0, dtype=ITYPE) # unused
         else:
-            predecessor_matrix = np.empty((0, N), dtype=ITYPE)
+            predecessor_matrix = np.empty((len(indices), 0), dtype=ITYPE)
+            source_matrix = np.empty((len(indices), 0), dtype=ITYPE) # unused
 
     if unweighted:
         csr_data = np.ones(csgraph.data.shape)
     else:
         csr_data = csgraph.data
-
+    
     if directed:
+        # for null transposed CSR
+        dummy_double_array = np.empty(0, dtype=DTYPE)
+        dummy_int_array = np.empty(0, dtype=ITYPE)
         if min_only:
-            _dijkstra_directed_multi(indices,
-                                     csr_data, csgraph.indices,
-                                     csgraph.indptr,
-                                     dist_matrix, predecessor_matrix,
-                                     source_matrix, limitf)
+            _dijkstra(indices,
+                      csr_data, csgraph.indices, csgraph.indptr,
+                      dummy_double_array, dummy_int_array, dummy_int_array,
+                      dist_matrix, predecessor_matrix, source_matrix,
+                      limitf)
         else:
-            _dijkstra_directed(indices,
-                               csr_data, csgraph.indices, csgraph.indptr,
-                               dist_matrix, predecessor_matrix, limitf)
+            _dijkstra_multi_separate(
+                      indices,
+                      csr_data, csgraph.indices, csgraph.indptr,
+                      dummy_double_array, dummy_int_array, dummy_int_array,
+                      dist_matrix, predecessor_matrix, source_matrix,
+                      limitf)
     else:
         csgraphT = csgraph.T.tocsr()
         if unweighted:
@@ -599,18 +607,18 @@ def dijkstra(csgraph, directed=True, indices=None,
         else:
             csrT_data = csgraphT.data
         if min_only:
-            _dijkstra_undirected_multi(indices,
-                                       csr_data, csgraph.indices,
-                                       csgraph.indptr,
-                                       csrT_data, csgraphT.indices,
-                                       csgraphT.indptr,
-                                       dist_matrix, predecessor_matrix,
-                                       source_matrix, limitf)
-        else:
-            _dijkstra_undirected(indices,
+            _dijkstra(indices,
                                  csr_data, csgraph.indices, csgraph.indptr,
                                  csrT_data, csgraphT.indices, csgraphT.indptr,
-                                 dist_matrix, predecessor_matrix, limitf)
+                                 dist_matrix, predecessor_matrix, source_matrix,
+                                 limitf)
+        else:
+            _dijkstra_multi_separate(
+                                 indices,
+                                 csr_data, csgraph.indices, csgraph.indptr,
+                                 csrT_data, csgraphT.indices, csgraphT.indptr,
+                                 dist_matrix, predecessor_matrix, source_matrix,
+                                 limitf)
 
     if return_predecessors:
         if min_only:
@@ -629,19 +637,20 @@ ctypedef pair[DTYPE_t, uint_t] dist_index_pair_t
 ctypedef priority_queue[dist_index_pair_t] dijkstra_queue_t
 
 @cython.boundscheck(False)
-cdef _dijkstra_scan_heap_multi(dijkstra_queue_t &heap,
+cdef _dijkstra_scan_heap(dijkstra_queue_t &heap,
                          dist_index_pair_t v,
                          const double[:] csr_weights,
                          const int[:] csr_indices,
                          const int[:] csr_indptr,
                          double[:] dist_matrix,
                          int[:] pred,
-                         int[:] sources,
                          int return_pred,
+                         int[:] sources,
+                         int return_source,
                          DTYPE_t limit):
     cdef:
-        unsigned int k, j_source, j_current
         ITYPE_t j
+        unsigned int j_current
         DTYPE_t next_val
 
     # v is a dist_index_pair_t poped from the queue
@@ -653,89 +662,24 @@ cdef _dijkstra_scan_heap_multi(dijkstra_queue_t &heap,
         if next_val <= limit:
             if dist_matrix[j_current] > next_val:
                 dist_matrix[j_current] = next_val
+                # The same vertex may be pushed multiple times to the queue, but
+                # anything with suboptimal distance is ignored when poped
                 heap.push(dist_index_pair_t(-next_val, j_current))
                 if return_pred:
                     pred[j_current] = v.second
+                if return_source:
                     sources[j_current] = sources[v.second]
 
-@cython.boundscheck(False)
-cdef _dijkstra_scan_heap(dijkstra_queue_t &heap,
-                         dist_index_pair_t v,
-                         const double[:] csr_weights,
-                         const int[:] csr_indices,
-                         const int[:] csr_indptr,
-                         double[:, :] dist_matrix,
-                         int[:, :] pred,
-                         int return_pred,
-                         DTYPE_t limit,
-                         int i):
-    cdef:
-        unsigned int j_current
-        ITYPE_t j
-        DTYPE_t next_val
-
-    # v is a dist_index_pair_t poped from the queue
-    # v.first: the distance of the vertex
-    # v.second: index of the vertex
-    for j in range(csr_indptr[v.second], csr_indptr[v.second + 1]):
-        j_current = csr_indices[j]
-        next_val = v.first + csr_weights[j]
-        if next_val <= limit:
-            if dist_matrix[i, j_current] > next_val:
-                dist_matrix[i, j_current] = next_val
-				# The same vertex may be pushed multiple times to the queue, but anything
-                heap.push(dist_index_pair_t(-next_val, j_current))
-                if return_pred:
-                    pred[i, j_current] = v.second
-
 
 @cython.boundscheck(False)
-cdef int _dijkstra_directed(
+cdef int _dijkstra(
             const int[:] source_indices,
             const double[:] csr_weights,
             const int[:] csr_indices,
             const int[:] csr_indptr,
-            double[:, :] dist_matrix,
-            int[:, :] pred,
-            DTYPE_t limit) except -1:
-    cdef:
-        unsigned int Nind = dist_matrix.shape[0]
-        unsigned int N = dist_matrix.shape[1]
-        unsigned int i, k, j_source, j_current
-        DTYPE_t next_val
-        int return_pred = (pred.size > 0)
-        dijkstra_queue_t heap # pairs of {-distance, vertex index} will be pushed to this priority queue
-        dist_index_pair_t v
-    
-    for i in range(Nind):
-        j_source = source_indices[i]
-        
-        dist_matrix[i, j_source] = 0
-        # priority_queue is max-heap, so negate the distance to make it min-heap
-        heap.push(dist_index_pair_t(-dist_matrix[i, j_source], j_source))
-
-        while heap.size():
-            v = heap.top()
-            heap.pop()
-            v.first = -v.first
-            # Do not process v if its distance has been updated
-			# after v was pushed to the queue, in which case
-			# _dijkstra_scan_heap should have already been called with
-			# the vertex v.second
-            # This assures _dijkstra_scan_heap is only called once per vertex
-			# and the total complexity is O(Mlog(M)) per source
-            if dist_matrix[i, v.second] < v.first : continue
-            
-            _dijkstra_scan_heap(heap, v, csr_weights, csr_indices, csr_indptr,
-                                dist_matrix, pred, return_pred, limit, i)
-    return 0
-
-@cython.boundscheck(False)
-cdef int _dijkstra_directed_multi(
-            const int[:] source_indices,
-            const double[:] csr_weights,
-            const int[:] csr_indices,
-            const int[:] csr_indptr,
+            const double[:] csrT_weights,
+            const int[:] csrT_indices,
+            const int[:] csrT_indptr,
             double[:] dist_matrix,
             int[:] pred,
             int[:] sources,
@@ -746,108 +690,74 @@ cdef int _dijkstra_directed_multi(
         unsigned int i, k, j_source, j_current
         DTYPE_t next_val
         int return_pred = (pred.size > 0)
-        dijkstra_queue_t heap
+        int return_sources = (sources.size > 0)
+        int directed = (csrT_weights.size == 0)
+        dijkstra_queue_t heap # pairs of {-distance, vertex index} will be pushed
         dist_index_pair_t v
-
+    
     for i in range(Nind):
         j_source = source_indices[i]
         dist_matrix[j_source] = 0
         heap.push(dist_index_pair_t(-dist_matrix[j_source], j_source))
-        if return_pred:
+        if return_sources:
             sources[j_source] = j_source
+    
+    if return_pred:
+        assert pred.size == N
+    if return_sources:
+        assert sources.size == N
     
     while heap.size():
         v = heap.top()
         heap.pop()
         v.first = -v.first
-        if dist_matrix[v.second] < v.first : continue
+        # Do not process v if its distance has been updated
+        # after v was pushed to the queue, in which case
+        # _dijkstra_scan_heap should have already been called with
+        # the vertex v.second
+        # This assures _dijkstra_scan_heap is only called once per vertex
+        # and the total complexity is O(Mlog(M)) per source
+        if dist_matrix[v.second] < v.first :
+            continue
         
-        _dijkstra_scan_heap_multi(heap, v, csr_weights, csr_indices, csr_indptr,
-                                  dist_matrix, pred, sources, return_pred, limit)
+        _dijkstra_scan_heap(heap, v, csr_weights, csr_indices, csr_indptr,
+                            dist_matrix, pred, return_pred,
+                            sources, return_sources, limit)
+        if not directed:
+            _dijkstra_scan_heap(heap, v,
+                                csrT_weights, csrT_indices, csrT_indptr,
+                                dist_matrix, pred, return_pred,
+                                sources, return_sources, limit)
     return 0
 
 @cython.boundscheck(False)
-cdef int _dijkstra_undirected(
-            int[:] source_indices,
-            double[:] csr_weights,
-            int[:] csr_indices,
-            int[:] csr_indptr,
-            double[:] csrT_weights,
-            int[:] csrT_indices,
-            int[:] csrT_indptr,
+cdef int _dijkstra_multi_separate(
+            const int[:] source_indices,
+            const double[:] csr_weights,
+            const int[:] csr_indices,
+            const int[:] csr_indptr,
+            const double[:] csrT_weights,
+            const int[:] csrT_indices,
+            const int[:] csrT_indptr,
             double[:, :] dist_matrix,
             int[:, :] pred,
-            DTYPE_t limit) except -1:
-    cdef:
-        unsigned int Nind = dist_matrix.shape[0]
-        unsigned int N = dist_matrix.shape[1]
-        unsigned int i, k, j_source, j_current
-        DTYPE_t next_val
-        int return_pred = (pred.size > 0)
-        dijkstra_queue_t heap
-        dist_index_pair_t v
-    
-    for i in range(Nind):
-        j_source = source_indices[i]
-        
-        dist_matrix[i, j_source] = 0
-        heap.push(dist_index_pair_t(-dist_matrix[i, j_source], j_source))
-
-        while heap.size():
-            v = heap.top()
-            heap.pop()
-            v.first = -v.first
-            if dist_matrix[i, v.second] < v.first : continue
-            
-            _dijkstra_scan_heap(heap, v, csr_weights, csr_indices, csr_indptr,
-                                dist_matrix, pred, return_pred, limit, i)
-            _dijkstra_scan_heap(heap, v, csrT_weights, csrT_indices, csrT_indptr,
-                                dist_matrix, pred, return_pred, limit, i)
-
-    return 0
-
-
-@cython.boundscheck(False)
-cdef int _dijkstra_undirected_multi(
-            int[:] source_indices,
-            double[:] csr_weights,
-            int[:] csr_indices,
-            int[:] csr_indptr,
-            double[:] csrT_weights,
-            int[:] csrT_indices,
-            int[:] csrT_indptr,
-            double[:] dist_matrix,
-            int[:] pred,
-            int[:] sources,
+            int[:, :] sources,
             DTYPE_t limit) except -1:
     cdef:
         unsigned int Nind = source_indices.shape[0]
-        unsigned int N = dist_matrix.shape[0]
-        unsigned int i, k, j_source, j_current
-        DTYPE_t next_val
-        int return_pred = (pred.size > 0)
-        dijkstra_queue_t heap
-        dist_index_pair_t v
-
+        unsigned int i
+        int source_list[1]
+    
+    assert dist_matrix.shape[0] == Nind
+    assert pred.shape[0] == Nind
+    assert sources.shape[0] == Nind
     for i in range(Nind):
-        j_source = source_indices[i]
-        dist_matrix[j_source] = 0
-        heap.push(dist_index_pair_t(-dist_matrix[j_source], j_source))
-        if return_pred:
-            sources[j_source] = j_source
-    
-    while heap.size():
-        v = heap.top()
-        heap.pop()
-        v.first = -v.first
-        if dist_matrix[v.second] < v.first : continue
-        
-        _dijkstra_scan_heap_multi(heap, v, csr_weights, csr_indices, csr_indptr,
-                                  dist_matrix, pred, sources, return_pred, limit)
-        _dijkstra_scan_heap_multi(heap, v, csrT_weights, csrT_indices, csrT_indptr,
-                                  dist_matrix, pred, sources, return_pred, limit)
-    
-    return 0
+        source_list[0] = source_indices[i]
+        _dijkstra(source_list,
+                  csr_weights, csr_indices, csr_indptr,
+                  csrT_weights, csrT_indices, csrT_indptr,
+                  dist_matrix[i], pred[i], sources[i], limit)
+
 
 
 def bellman_ford(csgraph, directed=True, indices=None,
@@ -1005,8 +915,7 @@ cdef int _bellman_ford_directed(
     cdef:
         unsigned int Nind = dist_matrix.shape[0]
         unsigned int N = dist_matrix.shape[1]
-        unsigned int i, j, j_source, count
-        int k
+        unsigned int i, j, k, j_source, count
         DTYPE_t d1, d2, w12
         int return_pred = (pred.size > 0)
 
@@ -1209,7 +1118,7 @@ def johnson(csgraph, directed=True, indices=None,
         predecessor_matrix = np.empty((len(indices), N), dtype=ITYPE)
         predecessor_matrix.fill(NULL_IDX)
     else:
-        predecessor_matrix = np.empty((0, N), dtype=ITYPE)
+        predecessor_matrix = np.empty((len(indices), 0), dtype=ITYPE)
 
     #------------------------------
     # initialize distance array
@@ -1235,19 +1144,27 @@ def johnson(csgraph, directed=True, indices=None,
     _johnson_add_weights(csr_data, csgraph.indices,
                          csgraph.indptr, dist_array)
 
+    dummy_source_matrix = np.empty((len(indices), 0), dtype=ITYPE)
     if directed:
-        _dijkstra_directed(indices,
-                           csr_data, csgraph.indices, csgraph.indptr,
-                           dist_matrix, predecessor_matrix, np.inf)
+        # for null transposed CSR
+        dummy_double_array = np.empty(0, dtype=DTYPE)
+        dummy_int_array = np.empty(0, dtype=ITYPE)
+        _dijkstra_multi_separate(
+            indices,
+            csr_data, csgraph.indices, csgraph.indptr,
+            dummy_double_array, dummy_int_array, dummy_int_array,
+            dist_matrix, predecessor_matrix, dummy_source_matrix, np.inf)
     else:
         csgraphT = csr_matrix((csr_data, csgraph.indices, csgraph.indptr),
                                csgraph.shape).T.tocsr()
         _johnson_add_weights(csgraphT.data, csgraphT.indices,
                              csgraphT.indptr, dist_array)
-        _dijkstra_undirected(indices,
-                             csr_data, csgraph.indices, csgraph.indptr,
-                             csgraphT.data, csgraphT.indices, csgraphT.indptr,
-                             dist_matrix, predecessor_matrix, np.inf)
+        _dijkstra_multi_separate(
+            indices,
+            csr_data, csgraph.indices, csgraph.indptr,
+            csgraphT.data, csgraphT.indices, csgraphT.indptr,
+            dist_matrix, predecessor_matrix, dummy_source_matrix,
+            np.inf)
 
     # ------------------------------
     # correct the distance matrix for the bellman-ford weights