Another pull request to speed up sparse.csgraph.dijkstra, succeeding #16696
Thank you @jjerphan for reviewing the previous pull request and giving me advice

 - Replaces FibonacciHeap with std::priority_queue which makes sparse.csgraph.dijkstra faster
      - asymptotically due to the removal of a bug in FibonacciHeap implementation
      - with regard to the constant factor due to better memory usage(?)
 - Unifies the four similar dijkstra functions into one
      - Performance overhead was negligible, probably because the priority_queue is the bottleneck
 - Adds the star graph to the benchmark as the worst case for the previous version
 - Fixes a typo in the test file name 
 - A part of the test file for dijkstra is modified for better readability, nothing really changed

Benchmark(i5-10600k, ArchLinux)
```
Before:
===== ========== ============= =============
--                          format          
---------------- ---------------------------
n    min_only      random         star    
===== ========== ============= =============
30     True      178±0.1μs     125±0.9μs  
30    False      191±0.4μs     131±0.8μs  
300     True      451±0.4μs     183±0.4μs  
300    False       4.78±0ms      1.67±0ms  
900     True     2.55±0.01ms     657±2μs   
900    False      102±0.08ms   48.4±0.08ms 
===== ========== ============= =============

After:
===== ========== ============= ===========
--                         format         
---------------- -------------------------
n    min_only      random        star   
===== ========== ============= ===========
30     True      178±0.1μs     126±1μs  
30    False      192±0.6μs    134±0.8μs 
300     True      363±0.4μs    152±0.3μs 
300    False       2.46±0ms     459±1μs  
900     True     1.93±0.01ms   200±0.2μs 
900    False     48.1±0.02ms    3.78±0ms 
===== ========== ============= ===========
```