@larsoner 2022-09-14
~3x more lines removed than added, code is more readable, uses standard library functions, and improves performance. Awesome!

@fcollman I remember you working on this some years ago in #9664, do you want to take a look?