@fcollman 2022-09-14
I looked through it and the logic seems sound to me, and agreed I like the redesign quite a bit.  It did make me look back through the tests and wonder if we couldn't come up with some more intentionally designed use cases of graphs as they seem a bit skimpy to me.  

For example, this introduces the star format graph which I like in the benchmarks, but the results aren't validated.  Perhaps some ground truth data generated using the old implementation would be good to add to the tests and show that things still work as we desire them to.