@tylerjereddy 2024-03-12
The review comments proper look pretty minor, but the comment at https://github.com/scipy/scipy/pull/17019#issuecomment-1247346395 raises some concerns about depth of testing here as we switch implementations. Given about a year of inactivity now, I'll bump the milestone again ahead of `1.13.0` branching soon-ish.