@tsery-ns 2024-05-15
> @tsery-ns feel free to open a new pr with these commits

PR https://github.com/scipy/scipy/pull/20717 ready for review 😄